pyproject.toml
[project]
name = "stripe-billinglakehouse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.1.7",
    "apache-airflow-providers-databricks>=7.13.0",
    "azure-identity>=1.25.3",
    "azure-keyvault-secrets>=4.11.0",
    "delta-spark>=4.1.0",
    "pyspark>=4.1.1",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "jobs*"]
